+++ src/GovUk.Frontend.AspNetCore/TagHelpers/TextInputContext.cs
@@ -4,12 +4,18 @@ namespace GovUk.Frontend.AspNetCore.TagHelpers;
 
 internal class TextInputContext : FormGroupContext3
 {
+    private (TemplateString Content, string TagName)? _beforeInput;
     private (InputOptionsPrefix Options, string TagName)? _prefix;
     private (InputOptionsSuffix Options, string TagName)? _suffix;
+    private (TemplateString Content, string TagName)? _afterInput;
+
+    public TemplateString? BeforeInput => _beforeInput?.Content;
 
     public InputOptionsPrefix? Prefix => _prefix?.Options;
 
     public InputOptionsSuffix? Suffix => _suffix?.Options;
+
+    public TemplateString? AfterInput => _afterInput?.Content;
 
     protected override IReadOnlyCollection<string> ErrorMessageTagNames { get; } =
         [/*TextInputTagHelper.ErrorMessageShortTagName, */TextInputErrorMessageTagHelper.TagName];
@@ -20,11 +26,17 @@ internal class TextInputContext : FormGroupContext3
     protected override IReadOnlyCollection<string> LabelTagNames { get; } =
         [/*TextInputTagHelper.LabelShortTagName, */TextInputLabelTagHelper.TagName];
 
+    private IReadOnlyCollection<string> BeforeInputTagNames { get; } =
+        [/*TextInputBeforeInputTagHelper.ShortTagName, */TextInputBeforeInputTagHelper.TagName];
+
     private IReadOnlyCollection<string> PrefixTagNames { get; } =
         [/*TextInputPrefixTagHelper.ShortTagName, */TextInputPrefixTagHelper.TagName];
 
     private IReadOnlyCollection<string> SuffixTagNames { get; } =
         [/*TextInputSuffixTagHelper.ShortTagName, */TextInputSuffixTagHelper.TagName];
+
+    private IReadOnlyCollection<string> AfterInputTagNames { get; } =
+        [/*TextInputAfterInputTagHelper.ShortTagName, */TextInputAfterInputTagHelper.TagName];
 
     protected override string RootTagName { get; } = TextInputTagHelper.TagName;
 
@@ -34,6 +46,13 @@ internal class TextInputContext : FormGroupContext3
         TemplateString? html,
         string tagName)
     {
+        if (_beforeInput is var (_, beforeInputTagName))
+        {
+            throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
+                tagName,
+                beforeInputTagName);
+        }
+
         if (_prefix is var (_, prefixTagName))
         {
             throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
@@ -46,6 +65,13 @@ internal class TextInputContext : FormGroupContext3
             throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
                 tagName,
                 suffixTagName);
+        }
+
+        if (_afterInput is var (_, afterInputTagName))
+        {
+            throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
+                tagName,
+                afterInputTagName);
         }
 
         base.SetErrorMessage(visuallyHiddenText, attributes, html, tagName);
@@ -56,6 +82,13 @@ internal class TextInputContext : FormGroupContext3
         TemplateString? html,
         string tagName)
     {
+        if (_beforeInput is var (_, beforeInputTagName))
+        {
+            throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
+                tagName,
+                beforeInputTagName);
+        }
+
         if (_prefix is var (_, prefixTagName))
         {
             throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
@@ -68,6 +101,13 @@ internal class TextInputContext : FormGroupContext3
             throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
                 tagName,
                 suffixTagName);
+        }
+
+        if (_afterInput is var (_, afterInputTagName))
+        {
+            throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
+                tagName,
+                afterInputTagName);
         }
 
         base.SetHint(attributes, html, tagName);
@@ -79,6 +119,13 @@ internal class TextInputContext : FormGroupContext3
         TemplateString? html,
         string tagName)
     {
+        if (_beforeInput is var (_, beforeInputTagName))
+        {
+            throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
+                tagName,
+                beforeInputTagName);
+        }
+
         if (_prefix is var (_, prefixTagName))
         {
             throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
@@ -93,7 +140,50 @@ internal class TextInputContext : FormGroupContext3
                 suffixTagName);
         }
 
+        if (_afterInput is var (_, afterInputTagName))
+        {
+            throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
+                tagName,
+                afterInputTagName);
+        }
+
         base.SetLabel(isPageHeading, attributes, html, tagName);
+    }
+
+    public void SetBeforeInput(TemplateString content, string tagName)
+    {
+        ArgumentNullException.ThrowIfNull(content);
+        ArgumentNullException.ThrowIfNull(tagName);
+
+        if (BeforeInput is not null)
+        {
+            throw ExceptionHelper.OnlyOneElementIsPermittedIn(
+                BeforeInputTagNames,
+                RootTagName);
+        }
+
+        if (_prefix is var (_, prefixTagName))
+        {
+            throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
+                tagName,
+                prefixTagName);
+        }
+
+        if (_suffix is var (_, suffixTagName))
+        {
+            throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
+                tagName,
+                suffixTagName);
+        }
+
+        if (_afterInput is var (_, afterInputTagName))
+        {
+            throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
+                tagName,
+                afterInputTagName);
+        }
+
+        _beforeInput = (content, tagName);
     }
 
     public void SetPrefix(InputOptionsPrefix options, string tagName)
@@ -115,6 +205,13 @@ internal class TextInputContext : FormGroupContext3
                 suffixTagName);
         }
 
+        if (_afterInput is var (_, afterInputTagName))
+        {
+            throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
+                tagName,
+                afterInputTagName);
+        }
+
         _prefix = (options, tagName);
     }
 
@@ -130,6 +227,28 @@ internal class TextInputContext : FormGroupContext3
                 RootTagName);
         }
 
+        if (_afterInput is var (_, afterInputTagName))
+        {
+            throw ExceptionHelper.ChildElementMustBeSpecifiedBefore(
+                tagName,
+                afterInputTagName);
+        }
+
         _suffix = (options, tagName);
+    }
+
+    public void SetAfterInput(TemplateString content, string tagName)
+    {
+        ArgumentNullException.ThrowIfNull(content);
+        ArgumentNullException.ThrowIfNull(tagName);
+
+        if (AfterInput is not null)
+        {
+            throw ExceptionHelper.OnlyOneElementIsPermittedIn(
+                AfterInputTagNames,
+                RootTagName);
+        }
+
+        _afterInput = (content, tagName);
     }
 }